+++ application/classes/Core/Application.php
@@ -15,152 +15,98 @@ namespace Core {
 		private $applicationPath;
 
 		/**
-		 * @var array Arguments provided through URI parts
-		 */
-		private $args;
-
-		/**
-		 * @var array Actual controller path routed by given args
+		 * @var ControllerName Actual controller path routed by given args
 		 */
 		private $executedControllerName;
 
 		/**
-		 * @var array Method name called on the master controller
+		 * @var MethodName Method name called on the master controller
 		 */
 		private $calledMethodName;
 
 		/**
-		 * @var \Core\Configuration Holds the Application-wide configuration object.
+		 * @var Router The router responsible for parsing request uri to callable system path
 		 */
-		private $iConfiguration;
+		private $router;
 
 		/**
 		 * Parse the current route and set caching as needed.
+		 * 
 		 * @param array $args Application arguments, usually url-parts divided by /, or argv.
 		 */
-		public function __construct(array $args) {
-			$this->applicationPath = APP;
-
-			$configurationFile = STORAGE . "/config/application.jsonc";
-
-			$this->iConfiguration = new Configuration($configurationFile);
-
-			\Resource::set(new \Database\Connection(
-				$this->iConfiguration->get("database.host"),
-				$this->iConfiguration->get("database.username"),
-				$this->iConfiguration->get("database.password"),
-				$this->iConfiguration->get("database.name")
-			));
-
-			if(CLI === false) {
-				$route = $args["route"] ?? $this->iConfiguration->get("default_route");
-				$this->args = explode('/', ltrim($route, '/'));
-			} else {
-				$this->args = array_slice($args, 1);
-			}
-		}
-
-		/**
-		 * Get an argument from the url. ommit $argIndex to get all arguments passed with the request.
-		 * This is set to a string because if the variable passed to this function 
-		 * is null it would be easier to debug with a ull rather than getting the whole array.
-		 * @param int $index the index or the url arg.
-		 * @return null|string, or null on failure.
-		 */
-		public function arg(int $index = -1) : ?string {
-			if(isset($this->args[$index]) && $this->args[$index] !== '') {
-				return $this->args[$index];
-			}
-
-			return null;
-		}
-
-		/**
-		 * Get all parsed application args
-		 * @return array
-		 */
-		public function getArgs() : array {
-			return $this->args;
+		public function __construct(Router $iRouter) {
+			$this->router = $iRouter;
 		}
 
 		/**
 		 * Get controller name/path of the executed main controller
-		 * @return string
+		 * 
+		 * @return ControllerName Called controller name
 		 */
-		public function getExecutedControllerName() : string {
+		public function getExecutedControllerName() : ControllerName {
 			return $this->executedControllerName;
 		}
 
 		/**
 		 * Get controller name/path of the executed main controller
-		 * @return string
+		 * 
+		 * @return MethodName Called method name
 		 */
-		public function getCalledMethodName() : string {
+		public function getCalledMethodName() : MethodName {
 			return $this->calledMethodName;
 		}
 
 		/**
-		 * Get the current working directory of the application
-		 * @return string
-		 */
-		public function getApplicationPath() : string {
-			return $this->applicationPath;
-		}
-
-		/**
-		 * Returns the configuration object associated with the application
-		 * @return Configuration - application-wide configuration
-		 */
-		public function getConfiguration() : Configuration {
-			return $this->iConfiguration;
-		}
-
-		/**
 		 * Get path to the specified controller file. Ommit the .php extension
+		 * 
 		 * @param string $controller name of the controller file.
 		 * @return string or null on failure
 		 */
 		public function getControllerPath(string $controller) : ?string {
-			return $this->getApplicationPath()."/controllers/".basename($controller).".php";
+			return APP_PATH."/controllers/".basename($controller).".php";
+		}
+
+		/**
+		 * Get the router being used
+		 */
+		public function getRouter() : Router {
+			return $this->router;
 		}
 
 		/**
 		 * Executes a given controller by name.
 		 * Reroutes to NotFouncController if a \Core\Exception\NotFound is thrown
 		 * within the controller or any of it's child controllers.
+		 * 
 		 * @param string $controller The controller name, alias the class name.
-		 * @return Controller The dispatched controller that has just been executed.
+		 * @return \Controller The dispatched controller that has just been executed.
 		 */
-		public function executeController(string $controllerClass, string $methodName = MethodName::DEFAULT, ?\Controller $parentController = null) : \Controller {
-			$controllerClass = (string) new ControllerName($controllerClass);
-			$methodName 	 = (string) new MethodName($methodName);
+		public function executeController(ControllerName $iControllerName, ?MethodName $iMethodName = MethodName::DEFAULT, ?\Controller $parentController = null) : \Controller {			
+			$controllerName = $iControllerName->toString();
+			$methodName = $iMethodName->toString();
 
-			if(method_exists($controllerClass, $methodName) !== true) {
-				$methodName = MethodName::DEFAULT;
-			}
-			
+			$this->executedControllerName = $iControllerName;
+			$this->calledMethodName = $iMethodName;
+
 			try {
-				$iController = new $controllerClass;
+				$iController = new $controllerName($this);
 
 				if($parentController !== null) {
 					$iController->setParent($parentController);
 					$iController->setData($parentController->getData());
-				} else {
-					$this->executedControllerName = $iController->getName();
-					$this->calledMethodName = $methodName;
 				}
-
+				
 				$iController->start();
 				$iController->$methodName();
 				$iController->stop();
 			} catch(\Core\Exception\Forbidden $e) {
-				$iController = $this->executeController("Forbidden");
+				$iController = $this->executeController(new ControllerName("Forbidden"));
 			} catch(\Core\Exception\NotFound $e) {
-				$iController = $this->executeController("NotFound");
+				$iController = $this->executeController(new ControllerName("NotFound"));
 			}
 
 			foreach($iController->getChildren() as $childControllerName) {
-				$childController = $this->executeController($childControllerName, MethodName::DEFAULT, $iController);
+				$childController = $this->executeController($childControllerName, new MethodName(MethodName::DEFAULT), $iController);
 				$iController->setData($childController->getData());
 			}
 
@@ -170,16 +116,11 @@ namespace Core {
 		/**
 		 * Dispatches a controller, based upon the requeted path.
 		 * Serves a NotfoundController if it doesn't exists
+		 * 
 		 * @return \Controller Instance of extended Controller
 		 */
 		public function run() : \Controller {
-			$controllerBase = $this->arg(0);
-
-			$arguments = [$controllerBase];
-
-			if($this->arg(1) !== null) $arguments[] = $this->arg(1);
-
-			return $this->executeController(...$arguments);
+			return $this->executeController(...$this->router->getRoute());
 		}
 	}
 }