+++ application/classes/Core/Exception/Logic.php
@@ -6,7 +6,7 @@ namespace Core\Exception {
 	 * @author Allan Thue Rehhoff
 	 */
 	class Logic extends \Exception {
-		public function __construct($message, $code = 0, Exception $previous = null) {
+		public function __construct($message, $code = 0, \Exception $previous = null) {
 			parent::__construct($message, $code, $previous);
 		}
 	}