+++ application/classes/Core/Theme.php
@@ -1,7 +1,5 @@
 <?php
 namespace Core {
-	use Resource;
-
 	/**
 	 * Loads and setups the current configured theme in use.
 	 *
@@ -14,7 +12,12 @@ namespace Core {
 		private $name;
 
 		/**
-		 * @var \Core\Configuration Holds the theme configuration object.
+		 * @var \Core\Assets The assets class
+		 */
+		private $assets;
+
+		/**
+		 * @var \Configuration Holds the theme configuration object.
 		 */
 		private $iConfiguration;
 
@@ -22,24 +25,29 @@ namespace Core {
 		 * Doesn't do much of interest, this shouldn't be required to mess with.
 		 * @return void
 		 */
-		public function __construct(string $themename) {
-			$this->name = $themename;
+		public function __construct(Assets $iAssets) {
+			$this->name = \Resource::getConfiguration()->get("theme");
+			$this->assets = $iAssets;
 
 			$configurationFile = STORAGE . "/config/" . $this->name . ".theme.jsonc";
-			$this->iConfiguration = (new Configuration($configurationFile));
+			$this->iConfiguration = (new \Configuration($configurationFile));
 
 			if($this->iConfiguration->get("version.version") == "@version") {
-				$this->iConfiguration->set("version.version", Resource::get("Core\Application")->getConfiguration()->get("version"));
+				$this->iConfiguration->set("version.version", \Resource::getConfiguration()->get("version"));
 			}
 
 			$this->addAssets();
 		}
 
+		public function getAssets() : Assets {
+			return $this->assets;
+		}
+
 		/**
-		 * Returns the configuration object associated with the application
-		 * @return \Core\Configuration - application-wide configuration
+		 * Returns the configuration object associated with the theme
+		 * @return \Configuration - application-wide theme configuration
 		 */
-		public function getConfiguration() : Configuration {
+		public function getConfiguration() : \Configuration {
 			return $this->iConfiguration;
 		}
 
@@ -48,7 +56,7 @@ namespace Core {
 		 * @return string
 		 */
 		public function getTemplatePath(string $tpl = '') : string {
-			$path = Resource::get("Core\Application")->getApplicationPath()."/themes/".$this->getName();
+			$path = APP_PATH."/themes/".$this->getName();
 
 			if($tpl != '') {
 				$path .= '/'.$tpl;
@@ -64,7 +72,7 @@ namespace Core {
 		 * @return string
 		 */
 		public function getDirectoryUri(string $file = '/') : string {
-			$protocol = SSL ? "https://" : "http://";
+			$protocol = IS_SSL ? "https://" : "http://";
 			$host  = isset($_SERVER["SERVER_NAME"]) ? $_SERVER["SERVER_NAME"] : "127.0.0.1";
 			$path = rtrim(dirname($_SERVER["PHP_SELF"]), "/\\");
 			$baseurl = $protocol.$host.$path;
@@ -87,7 +95,7 @@ namespace Core {
 		 */
 		private function maybeAddVersionNumber(string $url) : string {
 			if($this->iConfiguration->get("version.expose") === true) {
-				$proto = SSL ? "https://" : "http://";
+				$proto = IS_SSL ? "https://" : "http://";
 				$string = $proto . $_SERVER["HTTP_HOST"];
 
 				if(mb_substr($url, 0, mb_strlen($string)) === $string) {
@@ -120,7 +128,7 @@ namespace Core {
 
 						$src = $this->maybeAddVersionNumber($src);
 
-						Resource::get("Assets")->addJavascript($src);
+						$this->assets->addJavascript($src);
 					}
 				}
 			}
@@ -139,7 +147,7 @@ namespace Core {
 
 						$src = $this->maybeAddVersionNumber($src);
 
-						Resource::get("Assets")->addStylesheet($src, $region);
+						$this->assets->addStylesheet($src, $region);
 					}
 				}
 			}