+++ application/classes/Database/Connection.php
@@ -33,6 +33,9 @@
 			/** @var string Last query attempted to be executed. */
 			public $lastQuery;
 
+			/** @var int Number When an array is passed as criteria this will be incremented for each value across all arrays */
+			private $arrayINCounter = 0;
+
 			/**
 			* Initiate a new database connection using PDO as a driver.
 			*
@@ -195,10 +198,11 @@
 						if(is_array($filter) === true) {
 							$tmparr = [];
 
-							foreach ($filter as $i => $item) {
-								$key = "val".$i;
+							foreach($filter as $item) {
+								$key = "val".$this->arrayINCounter;
 								$tmparr[$key]  = $item;
 								$this->filters[$key] = $item;
+								$this->arrayINCounter++;
 							}
 
 							// (:val0, :val1, :val2)
@@ -319,12 +323,7 @@
 			*/
 			public function fetchCell(string $table, string $column, ?array $criteria = null) {
 				$sql = "SELECT `".$column."` FROM `".$table."` WHERE ".$this->keysToSql($criteria, "AND")." LIMIT 1";
-				$result = $this->query($sql, $criteria)->fetchColumn(0);
-
-				// Convert false values to null
-				// If we were to pass a false value to Entity::__construct();
-				// self::crateRow(); will end up trying to insert '0 => :0'
-				return $result !== false ? $result : null;
+				return $this->query($sql, $criteria)->fetchColumn(0);
 			}
 
 			/**
@@ -333,7 +332,7 @@
 			* @see \Database\Connection::fetchCell();
 			*/
 			public function fetchField(string $table, string $column, ?array $criteria = null) {
-				return $this->fetchCell($table, $column, $criteria);
+				return $this->fetchCell(...func_get_args());
 			}
 
 			/**
@@ -553,7 +552,7 @@
 				$statement = $this->prepare($query);
 
 				$tmpFilters = [];
-				foreach ($this->filters as $column => $value) {
+				foreach($this->filters as $column => $value) {
 					if(substr($column, 0, 1) !== ':') {
 						$arg = ":".$column;
 					} else {
@@ -566,6 +565,8 @@
 						$value = "'".$value."'";
 					} else if($type == "boolean") {
 						$value = $value ? 1 : 0;
+					} else if($type == "array") {
+						$value = "('" . implode("', '", $value) . "')";
 					}
 
 					$tmpFilters[$arg] = $value;