+++ application/classes/Database/Statement.php
@@ -5,7 +5,7 @@
 
 		class Statement extends PDOStatement {
 			protected function __construct(Connection $connection) {
-
+				// Dear future me, PDOStatement has to __construct() method
 			}
 
 			/**
@@ -19,6 +19,14 @@
 				// PHP < 8.0.0 compat. PDOStatement::fetchAll(); will return false
 				// if the result set was empty, fixed in PHP 8.0.0
 				return $result !== false ? $result : [];
+			}
+
+			public function fetchColumn(int $column = 0): mixed {
+				$result = parent::fetchColumn($column);
+
+				// PDOStatement::fetchColumn(); will return false
+				// We'll normalise it to return null
+				return $result !== false ? $result : null;
 			}
 		}
 	}