+++ application/classes/Resource.php
@@ -12,14 +12,16 @@
 		/**
 		 * Get an object by it's class name, namespaces included.
 		 * Instatiates the object if not already existing.
+		 * 
 		 * @return object on success
 		 */
 		public static function get(string $key) {
-			return (isset(self::$data[$key]) ? self::$data[$key] : NULL);
+			return self::$data[$key] ?? null;
 		}
 
 		/**
 		 * Stores a given object by it's class name, namespaces included
+		 * 
 		 * @return object The instance just stored. 
 		 */
 		public static function set($class) {
@@ -31,9 +33,28 @@
 
 		/**
 		 * Checks if the registry contains a given object.
+		 * 
 		 * @return bool
 		 */
 		public static function has(string $key) : bool {
 			return isset(self::$data[$key]);
+		}
+
+		/**
+		 * Get core application configuration object
+		 * 
+		 * @return \Core\Configuration
+		 */
+		public static function getConfiguration() {
+			return self::get("Configuration");
+		}
+
+		/**
+		 * Get instance of Database\Connection
+		 * 
+		 * @return \Database\Connection
+		 */
+		public static function getDatabaseConnection() : \Database\Connection {
+			return self::get("Database\Connection");
 		}
 	}